Icon support for FolderNode
